+++ src/VCRAccessories/CassetteExpirationException.php
@@ -9,7 +9,7 @@ class CassetteExpirationException extends \Exception
      *
      * @param string $message
      */
-    public function __construct($message = null)
+    public function __construct(string $message = '')
     {
         parent::__construct($message);
     }